requirements.txt
fastapi
uvicorn
xarray
pandas
psycopg2-binary
langchain
langchain-ollama
langchain-chroma
langchain-huggingface
huggingface-hub
sentence-transformers
indic-nlp-library
python-multipart